+++ src/main/java/com/baomidou/dynamic/datasource/annotation/Slave.java
@@ -14,31 +14,13 @@
  * limitations under the License.
  * <pre/>
  */
-package com.baomidou.dynamic.datasource.support;
+package com.baomidou.dynamic.datasource.annotation;
 
-/**
- * 动态数据源常量
- *
- * @author jobob
- * @since 2019-10-08
- */
-public interface DdConstants {
+import java.lang.annotation.*;
 
-    /**
-     * 数据源：主库
-     */
-    String MASTER = "master";
-    /**
-     * 数据源：从库
-     */
-    String SLAVE = "slave";
-
-    /**
-     * DRUID数据源类
-     */
-    String DRUID_DATASOURCE = "com.alibaba.druid.pool.DruidDataSource";
-    /**
-     * HikariCp数据源
-     */
-    String HIKARI_DATASOURCE = "com.zaxxer.hikari.HikariDataSource";
+@Target({ElementType.TYPE, ElementType.METHOD})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@DS("slave")
+public @interface Slave {
 }