+++ src/main/java/com/baomidou/dynamic/datasource/provider/YmlDynamicDataSourceProvider.java
@@ -31,12 +31,12 @@ import java.util.Map;
  */
 @Slf4j
 @AllArgsConstructor
-public class YmlDynamicDataSourceProvider extends AbstractDataSourceProvider implements DynamicDataSourceProvider {
+public class YmlDynamicDataSourceProvider extends AbstractDataSourceProvider {
 
     /**
      * 所有数据源
      */
-    private Map<String, DataSourceProperty> dataSourcePropertiesMap;
+    private final Map<String, DataSourceProperty> dataSourcePropertiesMap;
 
     @Override
     public Map<String, DataSource> loadDataSources() {